backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pypdf2==3.0.1
google-generativeai==0.3.2
python-jose==3.3.0
passlib==1.7.4
aiofiles==23.2.1
